# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/lifecycle/definitions": {
      "get": {
        "operationId": "LifecycleController_definitions",
        "summary": "Every state machine: states, transitions, roles and guards",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/lifecycle/{entityType}/{id}": {
      "get": {
        "operationId": "LifecycleController_inspect",
        "summary": "Current state, available transitions with their blockers, and history",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/lifecycle/{entityType}/{id}/transitions": {
      "post": {
        "operationId": "LifecycleController_fire",
        "summary": "Fire a transition. 409 with `blockers` if a guard stops it; senior officers may override overridable blockers with a reason.",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/objections": {
      "get": {
        "operationId": "ObjectionsController_list",
        "summary": "s.15 objections in the caller’s jurisdiction, with hearings",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "SUBMITTED",
                "UNDER_REVIEW",
                "HEARING_SCHEDULED",
                "RESOLVED",
                "REJECTED",
                "ESCALATED"
              ],
              "type": "string"
            }
          },
          {
            "name": "parcelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "objections"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ObjectionsController_file",
        "summary": "File an objection under s.15 (citizens: only for parcels they hold)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileObjectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "objections"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/objections/{id}/hearings": {
      "post": {
        "operationId": "ObjectionsController_scheduleHearing",
        "summary": "Schedule a hearing (moves the objection to HEARING_SCHEDULED)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleHearingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "objections"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/objections/hearings/{hearingId}/outcome": {
      "post": {
        "operationId": "ObjectionsController_recordOutcome",
        "summary": "Record that a hearing was held (with outcome notes) or adjourned",
        "parameters": [
          {
            "name": "hearingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HearingOutcomeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "objections"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/possession": {
      "get": {
        "operationId": "PossessionController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ELIGIBLE",
                "SCHEDULED",
                "POSSESSION_TAKEN",
                "HANDED_TO_PIA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "possession"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/possession/{parcelId}/take": {
      "post": {
        "operationId": "PossessionController_take",
        "summary": "Take possession of a parcel (s.38). 409 with blockers if payment, R&R or consent is incomplete.",
        "parameters": [
          {
            "name": "parcelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TakePossessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "possession"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/possession/{parcelId}/handover": {
      "post": {
        "operationId": "PossessionController_handover",
        "summary": "Hand the parcel over to the requiring body",
        "parameters": [
          {
            "name": "parcelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "possession"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "R-NLAM API",
    "description": "Real-Time National Land Acquisition & Management System. All routes need a Bearer token unless marked public. In AUTH_MODE=dev, get one from POST /api/auth/dev-login. Money is integer paise; dates are UTC.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TransitionDto": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string",
            "example": "COMPLETE_PAYMENT"
          },
          "reason": {
            "type": "string",
            "description": "Required (≥ 20 chars) when overriding a blocker"
          },
          "override": {
            "type": "boolean",
            "description": "Senior officers only: proceed despite overridable blockers"
          }
        },
        "required": [
          "event"
        ]
      },
...
      "FileObjectionDto": {
        "type": "object",
        "properties": {
          "parcelId": {
            "type": "string"
          },
          "applicant": {
            "type": "string",
            "example": "Ramesh Patil"
          },
          "category": {
            "type": "string",
            "enum": [
              "COMPENSATION_AMOUNT",
              "MEASUREMENT",
              "TITLE",
              "PUBLIC_PURPOSE",
              "RR_ENTITLEMENT",
              "OTHER"
            ]
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "parcelId",
          "applicant",
          "category",
          "description"
        ]
      },
      "ScheduleHearingDto": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "example": "2026-11-05T11:00:00+05:30"
          },
          "venue": {
            "type": "string",
            "example": "Collectorate Hall 2, Wardha"
          },
          "presidingOfficer": {
            "type": "string",
            "example": "Dy. Collector (LA), Wardha"
          }
        },
        "required": [
          "scheduledAt",
          "venue",
          "presidingOfficer"
        ]
      },
      "HearingOutcomeDto": {
        "type": "object",
        "properties": {
          "held": {
            "type": "boolean",
            "description": "true = hearing held; false = adjourned"
          },
          "outcome": {
            "type": "string"
          }
        },
        "required": [
          "held"
        ]
      },
...
      "TakePossessionDto": {
        "type": "object",
        "properties": {
          "takenOn": {
            "type": "string",
            "example": "2026-10-20"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          }
        }
      }
    }
  }
}